Parser/pyproject.toml
[project]
name = "sniplinebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fake-useragent>=2.2.0",
    "feedparser>=6.0.11",
    "httpx>=0.28.1",
    "lxml-html-clean>=0.4.1",
    "newspaper3k>=0.2.8",
    "redis>=5.2.1",
    "selenium>=4.31.0",
    "undetected-chromedriver>=3.5.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
]
